requirements.txt
# Source tag requirements:
# -r https://raw.githubusercontent.com/nasa/fprime/refs/tags/v4.0.0/requirements.txt

# Local requirements:
-r ./lib/fprime/requirements.txt
./lib/fprime-python/
pybind11==3.0.2
opencv-python==4.12.0.88
fprime_python_model